Lumberjack constants (#874)
* Added lumberJack constants to config

* Modified config data

* fixed tests

+++ cmd/vote.go
@@ -502,7 +502,7 @@ func (*UtilsStruct) GetLastProposedEpoch(client *ethclient.Client, blockNumber *
 	if err != nil {
 		return 0, err
 	}
-	bufferPercent, err := stringUtils.ParseInt(bufferPercentString)
+	bufferPercent, err := stringUtils.ParseInt64(bufferPercentString)
 	if err != nil {
 		return 0, err
 	}
